refactor(folder): Reorganized files in API folder for better structure Improve code maintainability and readability
Improve code maintainability and readability.

Api folder outside APP folder removed.
Changes in file names.
Changes in comments and some types added.
