+++ README.md
@@ -79,7 +79,7 @@ npm install
      - <a href='https://github.com/consumet/api.consumet.org'>Consumet API</a>
      - <a href='https://github.com/ghoshRitesh12/aniwatch-api'>Aniwatch API</a>
    - OPTIONAL: This project uses a JSON file (47 mb) filled with Animes and Mangas data as a offline Database. This repository already has this file, but it might be outdated, so you decide if you want to ignore this step.
-     - Go to <a href='https://github.com/manami-project/anime-offline-database'>this repository</a> and download the JSON file that will be used on only `Search Page` (or you can make some changes and use some API to fetch the data).
+     - Go to <a href='https://github.com/manami-project/anime-offline-database'>anime-offline-database</a> and download the JSON file that will be used on only `Search Page` (or you can make some changes and use some API to fetch the data).
      - With the file downloaded, put it in the `/app/api/anime-info` directory, replacing the previous one.
    - On Firebase, get your configs to use the Authentication and Firestore Database.
      - All the Firebase info needed bellow can be found when you create a new project.
@@ -90,8 +90,8 @@ With all that done, you will need to fill the `.env.local` like the example bell
 ```javascript
 NEXT_PUBLIC_CONSUMET_API_URL=https://your-hosted-consumet-api-url.com
 NEXT_PUBLIC_ANIWATCH_API_URL=https://your-hosted-aniwatch-api-url.com
-// Make sure to add the pathname "/api/anime-info" bellow
-NEXT_PUBLIC_NEXT_INTERNAL_API_URL=https://url-to-where-your-website-is-hosted.com/api/anime-info
+// Make sure to add the pathname "/api/animes-database" bellow
+NEXT_PUBLIC_NEXT_INTERNAL_API_URL=https://url-to-where-your-website-is-hosted.com/api/animes-database
 NEXT_PUBLIC_FIREBASE_API_KEY=firebase-setting-related-to-this-field
 NEXT_PUBLIC_FIREBASE_AUTH_DOMAIN=firebase-setting-related-to-this-field
 NEXT_PUBLIC_FIREBASE_PROJECT_ID=firebase-setting-related-to-this-field
