+++ app/api/anilistQueryFunctions.ts
