+++ app/components/HomePage/NavThoughMedias/index.tsx
@@ -3,7 +3,7 @@ import React, { SetStateAction, useEffect, useState } from 'react'
 import styles from './component.module.css'
 import Link from 'next/link'
 import { ApiAiringMidiaResults, ApiDefaultResult } from '@/app/ts/interfaces/apiAnilistDataInterface'
-import anilist from '@/api/anilist'
+import anilist from '@/app/api/anilist'
 import ChevronLeftIcon from '@/public/assets/chevron-left.svg'
 import ChevronRightIcon from '@/public/assets/chevron-right.svg'
 import CloseSvg from '@/public/assets/x.svg'
