+++ app/components/HomePage/NewestMediaSection/index.tsx
@@ -5,7 +5,7 @@ import MediaListCoverInfo2 from '../../MediaItemCoverInfo2'
 import CardMediaCoverAndDescription from '../../CardMediaCoverAndDescription'
 import NavButtons from '../../NavButtons'
 import { ApiAiringMidiaResults, ApiDefaultResult } from '@/app/ts/interfaces/apiAnilistDataInterface'
-import anilist from "@/api/anilist"
+import anilist from "@/app/api/anilist"
 import { useAuthState } from 'react-firebase-hooks/auth'
 import { getAuth } from 'firebase/auth'
 import { initFirebase } from '@/app/firebaseApp'
