+++ app/lib/fetchAnimeOnApi.ts
@@ -1,7 +1,7 @@
-import aniwatch from "@/api/aniwatch"
+import aniwatch from "@/app/api/aniwatch"
 import { EpisodesFetchedAnimeWatch, MediaInfoAniwatch } from "../ts/interfaces/apiAnimewatchInterface"
 import simulateRange from "./simulateRange"
-import gogoanime from "@/api/gogoanime"
+import gogoanime from "@/app/api/consumetGoGoAnime"
 import { MediaEpisodes, MediaInfo, MediaSearchResult } from "../ts/interfaces/apiGogoanimeDataInterface"
 import stringToOnlyAlphabetic from "./convertStringsTo"
 import { checkAnilistTitleMisspelling } from "./checkApiMediaMisspelling"
