+++ app/ts/interfaces/apiImdbInterface.d.ts
@@ -12,7 +12,6 @@ export interface ImdbSearchItem {
 export interface ImdbMediaInfo {
 
     rating: number,
-    vidsrcId: number,
     id: string,
     title: string,
     type: string,
